+++ src/1_1/src/silica.h
@@ -1,3 +1,6 @@
+// Common definitions for SiliCa
+// JIS X 6319-4 compatible card implementation
+
 #pragma once
 #include <stdint.h>
 
@@ -7,14 +10,7 @@ typedef const uint8_t *packet_t;
 
 // Functions for serial output
 // Similar to Arduino interface
-void Serial_write(uint8_t);
-void Serial_print(const char *);
-void Serial_println(const char *);
-
-// application layer functions
-void initialize();
-packet_t process(packet_t);
+void Serial_write(uint8_t data);
+void Serial_print(const char *str);
+void Serial_println(const char *str);
 void save_error(packet_t);
-
-// debug functions
-void print_packet(packet_t);
